requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
Pillow==5.2.0
requests==2.19.1
urllib3==1.23
docopt==0.6.2
